requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
bidict==0.21.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dateparser==1.0.0
dnspython==1.16.0
Flask==2.0.1
Flask-SocketIO==5.1.0
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.0.4
h11==0.12.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
python-binance==1.0.12
python-dateutil==2.8.1
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
simple-websocket==0.2.0
six==1.16.0
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.5
uWSGI==2.0.19.1
websockets==9.1
Werkzeug==2.0.1
wsproto==1.0.0
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0
